+++ lib/mdex/document.ex
@@ -875,7 +875,10 @@ end
 
 defmodule MDEx.Alert do
   @moduledoc """
-  GitHub style alerts / admonitions.
+  GitHub and GitLab style alerts / admonitions.
+
+  See [GitHub](https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#alerts)
+  and [GitLab](https://docs.gitlab.com/user/markdown/#alerts) docs.
   """
 
   @type t :: %__MODULE__{